fix respiratory structure.
