+++ .github/workflows/pr-checks.yml
@@ -9,7 +9,7 @@ jobs:
       - uses: actions/checkout@v3
 
       - name: Run tests
-        run: kotlinc Personal-Finance-Tracker/src/main/kotlin/datasource/*.kt Personal-Finance-Tracker/src/main/kotlin/models/*.kt Personal-Finance-Tracker/src/main/kotlin/test/*.kt -include-runtime -d test.jar
+        run: kotlinc src/main/kotlin/datasource/*.kt src/main/kotlin/models/*.kt src/main/kotlin/test/*.kt -include-runtime -d test.jar
 
       - name: Run tests Jar
         run:  java -jar test.jar || exit 1
