--- Personal-Finance-Tracker/.github/workflows/pr-checks.yml
@@ -1,15 +1,0 @@
-name: Kotlin PR Checks
-
-on: [pull_request]
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Run tests
-        run: kotlinc src/main/kotlin/datasource/*.kt src/main/kotlin/models/*.kt src/main/kotlin/test/*.kt -include-runtime -d test.jar
-
-      - name: Run tests Jar
-        run:  java -jar test.jar || exit 1
