--- Personal-Finance-Tracker/build.gradle.kts
@@ -1,21 +1,0 @@
-plugins {
-    kotlin("jvm") version "2.1.10"
-}
-
-group = "org.example"
-version = "1.0-SNAPSHOT"
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation(kotlin("test"))
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-kotlin {
-    jvmToolchain(17)
-}
