--- Personal-Finance-Tracker/settings.gradle.kts
@@ -1,5 +1,0 @@
-plugins {
-    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
-}
-rootProject.name = "Personal-Finance-Tracker"
-
