--- Personal-Finance-Tracker/src/main/kotlin/Main.kt
@@ -1,5 +1,0 @@
-package org.example
-
-fun main() {
-
-}
