--- Personal-Finance-Tracker/src/main/kotlin/models/Category.kt
@@ -1,9 +1,0 @@
-package models
-
-enum class Category {
-    FOOD,
-    RENT,
-    SALARY,
-    ENTERTAINMENT,
-    TRANSPORT
-}
