--- Personal-Finance-Tracker/src/main/kotlin/test/TransactionCreationTest.kt
@@ -1,2 +1,0 @@
-package org.example.test
-
