+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Apr 08 07:56:47 EET 2025
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.10-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
