+++ settings.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "0.8.0"
+}
+rootProject.name = "Personal-Finance-Tracker"
+
