+++ src/main/kotlin/Main.kt
@@ -1,0 +1,5 @@
+package org.example
+
+fun main() {
+
+}
