+++ src/main/kotlin/models/Category.kt
@@ -1,0 +1,9 @@
+package models
+
+enum class Category {
+    FOOD,
+    RENT,
+    SALARY,
+    ENTERTAINMENT,
+    TRANSPORT
+}
