+++ transaction.txt
@@ -1,0 +1,1 @@
+[{"id":"636ca86c-83d6-4bb2-a16f-bce0f92d75b9","name":"Name of the list 1","isDeposit":false,"amount":23.5,"category":"RENT","creationTime":"1999-08-09","editTime":[]},{"id":"43cf950b-3b89-44df-9b57-fa71e42c2f42","name":"Name of the list 2","isDeposit":false,"amount":47.0,"category":"SALARY","creationTime":"1999-08-09","editTime":[]},{"id":"233b2e2c-4be4-49ee-a1ab-6739fbf47a11","name":"Name of the list 3","isDeposit":false,"amount":70.5,"category":"TRANSPORT","creationTime":"1999-08-09","editTime":[]},{"id":"90789934-a829-41d8-a273-6d4b6d7fca79","name":"Name of the list 4","isDeposit":false,"amount":94.0,"category":"FOOD","creationTime":"1999-08-09","editTime":[]},{"id":"787c5fb5-1d02-4eb3-916b-bbe0298194cb","name":"Name of the list 5","isDeposit":false,"amount":117.5,"category":"ENTERTAINMENT","creationTime":"1999-08-09","editTime":[]}]
